requirements.txt
Flask==1.0.2
line-bot-sdk==1.15.0
requests==2.22.0
beautifulsoup4==4.8.0
gunicorn